requirements.txt
appdirs==1.4.4
asgiref==3.3.4
awsebcli==3.19.3
beautifulsoup4==4.9.1
bootstrap4==0.1.0
botocore==1.19.47
cement==2.8.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
cli-helpers==2.1.0
click==7.1.2
colorama==0.4.3
configobj==5.0.6
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.2.2
django-allauth==0.42.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-mapbox-location-field==1.6.1
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==4.3.0
filelock==3.0.12
future==0.16.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
Markdown==3.3.4
mysqlclient==2.0.3
oauthlib==3.1.0
pathspec==0.5.9
paypalhttp==1.0.0
pendulum==2.1.2
pgcli==3.1.0
pgspecial==1.12.1
Pillow==8.1.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pypiwin32==223
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pytzdata==2020.1
pywin32==300
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
self==2020.12.3
semantic-version==2.5.0
setproctitle==1.2.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tabulate==0.8.9
termcolor==1.1.0
terminaltables==3.1.0
urllib3==1.25.9
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
zipp==3.1.0
django-environ==0.4.5
